add log on folder service

+++ src/main/java/com/aisip/OnO/backend/service/FolderServiceImpl.java
@@ -6,16 +6,20 @@ import com.aisip.OnO.backend.Dto.Problem.ProblemResponseDto;
 import com.aisip.OnO.backend.entity.Folder;
 import com.aisip.OnO.backend.entity.Problem;
 import com.aisip.OnO.backend.entity.User.User;
+import com.aisip.OnO.backend.exception.FolderNotFoundException;
+import com.aisip.OnO.backend.exception.UserNotFoundException;
 import com.aisip.OnO.backend.repository.FolderRepository;
 import com.aisip.OnO.backend.repository.ProblemRepository;
 import com.aisip.OnO.backend.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 @Transactional
@@ -34,8 +38,10 @@ public class FolderServiceImpl implements FolderService {
         Optional<Folder> optionalRootFolder = folderRepository.findByUserIdAndParentFolderIsNull(userId);
 
         if (optionalRootFolder.isPresent()) {
+            log.info("find root folder id: " + optionalRootFolder.get().getId());
             return findFolder(userId, optionalRootFolder.get().getId());
         } else {
+            log.info("create root folder for userId: " + userId);
             return createRootFolder(userId, "메인");
         }
     }
@@ -44,7 +50,7 @@ public class FolderServiceImpl implements FolderService {
     public FolderResponseDto createRootFolder(Long userId, String folderName) {
         Optional<User> optionalUser = userRepository.findById(userId);
         if (optionalUser.isPresent()) {
-            Folder rootFolder = Folder.builder().name(folderName).user(optionalUser.get()).parentFolder(null).build();
+            Folder rootFolder = Folder.builder().name(folderName).user(optionalUser.get()).build();
 
             folderRepository.save(rootFolder);
 
@@ -58,14 +64,17 @@ public class FolderServiceImpl implements FolderService {
             return findFolder(userId, rootFolder.getId());
         }
 
-        return null;
+        throw new UserNotFoundException("유저를 찾을 수 없습니다!");
     }
 
     @Override
     public FolderResponseDto createFolder(Long userId, String folderName, Long parentFolderId) {
         Optional<User> optionalUser = userRepository.findById(userId);
         if (optionalUser.isPresent()) {
-            Folder folder = Folder.builder().name(folderName).user(optionalUser.get()).build();
+            Folder folder = Folder.builder()
+                    .name(folderName)
+                    .user(optionalUser.get())
+                    .build();
 
             if (parentFolderId != null) {
                 Optional<Folder> parentFolder = folderRepository.findById(parentFolderId);
@@ -77,7 +86,7 @@ public class FolderServiceImpl implements FolderService {
             return findFolder(userId, folder.getId());
         }
 
-        return null;
+        throw new UserNotFoundException("유저를 찾을 수 없습니다!");
     }
 
     @Override
@@ -89,18 +98,41 @@ public class FolderServiceImpl implements FolderService {
 
             Folder folder = optionalFolder.get();
 
-            FolderThumbnailResponseDto parentFolder = FolderThumbnailResponseDto.builder().folderId(folder.getParentFolder().getId()).folderName(folder.getParentFolder().getName()).build();
+            FolderThumbnailResponseDto parentFolder = null;
+            if (folder.getParentFolder() != null) {
+                parentFolder = FolderThumbnailResponseDto.builder()
+                        .folderId(folder.getParentFolder().getId())
+                        .folderName(folder.getParentFolder().getName())
+                        .build();
+            }
 
-            List<FolderThumbnailResponseDto> subFolders = folder.getSubFolders().stream().map(subFolder -> {
-                return FolderThumbnailResponseDto.builder().folderId(subFolder.getId()).folderName(subFolder.getName()).build();
-            }).toList();
+
+            List<FolderThumbnailResponseDto> subFolders = null;
+
+            if (folder.getSubFolders() != null) {
+                subFolders = folder.getSubFolders().stream().map(subFolder -> {
+                    return FolderThumbnailResponseDto.builder()
+                            .folderId(subFolder.getId())
+                            .folderName(subFolder.getName())
+                            .build();
+
+                }).toList();
+            }
 
             List<ProblemResponseDto> problems = problemService.findAllProblemsByFolderId(folderId);
 
-            return FolderResponseDto.builder().folderId(folder.getId()).folderName(folder.getName()).parentFolder(parentFolder).subFolders(subFolders).problems(problems).updateAt(folder.getUpdatedAt()).createdAt(folder.getCreatedAt()).build();
+            return FolderResponseDto.builder()
+                    .folderId(folder.getId())
+                    .folderName(folder.getName())
+                    .parentFolder(parentFolder)
+                    .subFolders(subFolders)
+                    .problems(problems)
+                    .updateAt(folder.getUpdatedAt())
+                    .createdAt(folder.getCreatedAt())
+                    .build();
         }
 
-        return null;
+        throw new FolderNotFoundException("폴더를 찾을 수 없습니다!");
     }
 
     @Override
@@ -124,12 +156,19 @@ public class FolderServiceImpl implements FolderService {
             }
         }
 
-        return null;
+        throw new UserNotFoundException("유저를 찾을 수 없습니다!");
     }
 
     @Override
-    public void deleteFolder(Long folderId) {
-        folderRepository.deleteById(folderId);
+    public void deleteFolder(Long userId, Long folderId) {
+
+        Optional<Folder> optionalFolder = folderRepository.findById(folderId);
+
+        if (optionalFolder.isPresent() && optionalFolder.get().getUser().getId().equals(userId)) {
+            folderRepository.deleteById(folderId);
+        } else{
+            throw new FolderNotFoundException("폴더를 찾을 수 없습니다!");
+        }
     }
 
     @Override
@@ -137,8 +176,15 @@ public class FolderServiceImpl implements FolderService {
 
         List<Folder> folders = folderRepository.findAllByUserId(userId);
 
-        return folders.stream().map(folder -> {
-            return FolderThumbnailResponseDto.builder().folderId(folder.getId()).folderName(folder.getName()).build();
-        }).toList();
+        if(folders != null){
+            return folders.stream().map(folder -> {
+                return FolderThumbnailResponseDto.builder()
+                        .folderId(folder.getId())
+                        .folderName(folder.getName())
+                        .build();
+            }).toList();
+        }
+
+        return null;
     }
 }
